+++ qc2/algorithms/pennylane/vqe/sa_oo_vqe.py
@@ -7,10 +7,10 @@ from qc2.algorithms.pennylane.vqe.vqe import VQE
 from qc2.algorithms.utils.orbital_optimization import OrbitalOptimization
 from qc2.algorithms.results import SAOOVQEResults
 from qc2.algorithms.pennylane.ansatz.state_resolution import state_resolution_initializer
-from qc2.second_q.fermionic_operator import FermionicOperator
+from qc2.algorithms.second_q.fermionic_operator import FermionicOperator
 from qc2.algorithms.pennylane.ansatz.generate_ansatz import generate_state_resolution_ansatz
 from qc2.qc2_driver import QC2
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.pennylane.qubit_mappers.base_mapper import PennylaneBaseMapper
 
 class SA_OO_VQE(VQE):
